refactor: controller 기능을 하는 RacingGameController 추가
